requirements.txt
streamlit
yfinance
pandas
numpy
plotly
pytest
openai
python-dotenv
openai[agents]
openai-agents
